requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.2.0
protobuf==4.21.12
python-dotenv==1.0.0
